# p2p/rlpx: reduce allocation and syscalls