@holiman 2021-05-21
Some charts from a snap-sync, where this PR is yellow, `master` is green. 
![Screenshot_2021-05-21 Dual Geth - Grafana(3)](https://user-images.githubusercontent.com/142290/119102662-906e5d00-ba1a-11eb-9862-45b09b10c149.png)
![Screenshot_2021-05-21 Dual Geth - Grafana(2)](https://user-images.githubusercontent.com/142290/119102666-9106f380-ba1a-11eb-8f31-61ebb7de994f.png)
![Screenshot_2021-05-21 Dual Geth - Grafana(1)](https://user-images.githubusercontent.com/142290/119102671-9106f380-ba1a-11eb-828c-428a22c40e0a.png)
![Screenshot_2021-05-21 Dual Geth - Grafana](https://user-images.githubusercontent.com/142290/119102672-919f8a00-ba1a-11eb-8c41-8ee894341302.png)


No really visible effects, but nothing broken, so LGTM