@fjl 2021-05-21
I'm also still testing this, and it seems the snappy buffer management doesn't fully work yet.