@holiman 2021-05-25
The `computeFrame` function calculates the hash, then calls `m.comput` which does it all over again, costing an extra alloc. If that is removed: 
```
[user@work rlpx]$ benchstat before.txt after1.txt 
name          old time/op    new time/op    delta
Throughput-6    18.4µs ±12%    18.4µs ±13%     ~     (p=0.573 n=8+10)

name          old speed      new speed      delta
Throughput-6  54.7MB/s ± 4%  55.9MB/s ±14%     ~     (p=0.315 n=7+10)

name          old alloc/op   new alloc/op   delta
Throughput-6    4.81kB ± 0%    3.84kB ± 0%  -20.08%  (p=0.000 n=8+8)

name          old allocs/op  new allocs/op  delta
Throughput-6      20.0 ± 0%      16.0 ± 0%  -20.00%  (p=0.000 n=10+10)

```
Also, unrolling the loop in that new `compute` function had a non-negligible impact, it seems: 
```
[user@work rlpx]$ benchstat before.txt after2.txt 
name          old time/op    new time/op    delta
Throughput-6    18.4µs ±12%    17.0µs ± 6%   -7.55%  (p=0.008 n=8+9)

name          old speed      new speed      delta
Throughput-6  54.7MB/s ± 4%  60.3MB/s ± 6%  +10.20%  (p=0.000 n=7+9)

name          old alloc/op   new alloc/op   delta
Throughput-6    4.81kB ± 0%    3.84kB ± 0%  -20.07%  (p=0.000 n=8+8)

name          old allocs/op  new allocs/op  delta
Throughput-6      20.0 ± 0%      16.0 ± 0%  -20.00%  (p=0.000 n=10+10)
```

Hacks: 
```diff
diff --git a/p2p/rlpx/rlpx.go b/p2p/rlpx/rlpx.go
index 72ba28d2ca..9132129ea9 100644
--- a/p2p/rlpx/rlpx.go
+++ b/p2p/rlpx/rlpx.go
@@ -272,7 +272,6 @@ func (m *hashMAC) compute(seed []byte) []byte {
 	if len(seed) != len(m.aesBuffer) {
 		panic("invalid MAC seed")
 	}
-
 	sum1 := m.hash.Sum(m.hashBuffer[:0])
 	m.cipher.Encrypt(m.aesBuffer[:], sum1)
 	for i := range m.aesBuffer {
@@ -283,11 +282,42 @@ func (m *hashMAC) compute(seed []byte) []byte {
 	return sum2[:16]
 }
 
+func (m *hashMAC) compute2(seed []byte) []byte {
+	if len(seed) != len(m.aesBuffer) {
+		panic("invalid MAC seed")
+	}
+	copy(m.hashBuffer[:0], seed)
+	sum1 := m.hashBuffer[0:16]
+	m.cipher.Encrypt(m.aesBuffer[:], sum1)
+
+	m.aesBuffer[0] ^= seed[0]
+	m.aesBuffer[1] ^= seed[1]
+	m.aesBuffer[2] ^= seed[2]
+	m.aesBuffer[3] ^= seed[3]
+	m.aesBuffer[4] ^= seed[4]
+	m.aesBuffer[5] ^= seed[5]
+	m.aesBuffer[6] ^= seed[6]
+	m.aesBuffer[7] ^= seed[7]
+	m.aesBuffer[8] ^= seed[8]
+	m.aesBuffer[9] ^= seed[9]
+	m.aesBuffer[10] ^= seed[10]
+	m.aesBuffer[11] ^= seed[11]
+	m.aesBuffer[12] ^= seed[12]
+	m.aesBuffer[13] ^= seed[13]
+	m.aesBuffer[14] ^= seed[14]
+	m.aesBuffer[15] ^= seed[15]
+
+	m.hash.Write(m.aesBuffer[:])
+
+	sum2 := m.hash.Sum(m.hashBuffer[:0])
+	return sum2[:16]
+}
+
 // computeFrame computes the MAC of framedata.
 func (m *hashMAC) computeFrame(framedata []byte) []byte {
 	m.hash.Write(framedata)
 	seed := m.hash.Sum(m.seedBuffer[:0])
-	return m.compute(seed[:16])
+	return m.compute2(seed[:16])
 }
 
 // Handshake performs the handshake. This must be called before any data is written
```