@holiman 2021-05-26
Current status: 

50% is sha3:
```
   20420ms 34.94% 34.94%    20420ms 34.94%  golang.org/x/crypto/sha3.keccakF1600
     970ms  1.66% 75.07%     8590ms 14.70%  golang.org/x/crypto/sha3.(*state).padAndPermute
```
20% syscalls: 
 ```
  11890ms 20.34% 55.28%    12280ms 21.01%  syscall.Syscall
```
20% AES:
```
    4990ms  8.54% 63.82%     4990ms  8.54%  crypto/aes.encryptBlockAsm
    1040ms  1.78% 73.41%     6550ms 11.21%  crypto/cipher.(*ctr).refill
```
Remaining 10% is mainly test-overhead: 
```
    3240ms  5.54% 69.36%     4220ms  7.22%  runtime.suspendG
    1330ms  2.28% 71.63%     1370ms  2.34%  runtime.nanotime (inline)
     720ms  1.23% 76.30%      720ms  1.23%  runtime.memclrNoHeapPointers
     700ms  1.20% 77.50%      840ms  1.44%  github.com/golang/snappy.encodeBlock
     700ms  1.20% 78.70%      700ms  1.20%  runtime.memmove
```